int64

+++ tg_api_client.go
@@ -120,7 +120,7 @@ func (c *TgApiClient) startConn(host string, port int) {
 	}
 }
 
-func (c *TgApiClient) sendMessage(botName string, chatId int64, text string, replyMsgId int32) (string, error) {
+func (c *TgApiClient) sendMessage(botName string, chatId int64, text string, replyMsgId int64) (string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5) // 5초 타임아웃
 	defer cancel()
 	r, err := c.client.SendMessage(ctx, &pb.SendMessageRequest{
@@ -135,7 +135,7 @@ func (c *TgApiClient) sendMessage(botName string, chatId int64, text string, rep
 	return r.GetMessage(), nil
 }
 
-func (c *TgApiClient) sendKeyboard(botName string, chatId int64, text string, buttons []*pb.InlineKeyboardData, rowNum int32, replyMsgId int32) (string, error) {
+func (c *TgApiClient) sendKeyboard(botName string, chatId int64, text string, buttons []*pb.InlineKeyboardData, rowNum int32, replyMsgId int64) (string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 	defer cancel()
 	r, err := c.client.SendKeyboard(ctx, &pb.SendKeyboardRequest{
@@ -152,7 +152,7 @@ func (c *TgApiClient) sendKeyboard(botName string, chatId int64, text string, bu
 	return r.GetMessage(), nil
 }
 
-func (c *TgApiClient) SendReplyInput(botName string, chatId int64, messageId int32, text string) (string, error) {
+func (c *TgApiClient) SendReplyInput(botName string, chatId int64, messageId int64, text string) (string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 	defer cancel()
 
@@ -173,7 +173,7 @@ func (c *TgApiClient) SendText(botName string, chatId int64, text string) (strin
 	return c.sendMessage(botName, chatId, text, 0)
 }
 
-func (c *TgApiClient) SendReplyText(botName string, chatId int64, replyMsgId int32, text string) (string, error) {
+func (c *TgApiClient) SendReplyText(botName string, chatId int64, replyMsgId int64, text string) (string, error) {
 	return c.sendMessage(botName, chatId, text, replyMsgId)
 }
 
@@ -181,11 +181,11 @@ func (c *TgApiClient) SendKeyboard(botName string, chatId int64, text string, bu
 	return c.sendKeyboard(botName, chatId, text, buttons, rowNum, 0)
 }
 
-func (c *TgApiClient) SendReplyKeyboard(botName string, chatId int64, replyMsgId int32, text string, buttons []*pb.InlineKeyboardData, rowNum int32) (string, error) {
+func (c *TgApiClient) SendReplyKeyboard(botName string, chatId int64, replyMsgId int64, text string, buttons []*pb.InlineKeyboardData, rowNum int32) (string, error) {
 	return c.sendKeyboard(botName, chatId, text, buttons, rowNum, replyMsgId)
 }
 
-func (c *TgApiClient) SendFile(botName string, chatId int64, filePath string, caption string, replyMsgId int32) (string, error) {
+func (c *TgApiClient) SendFile(botName string, chatId int64, filePath string, caption string, replyMsgId int64) (string, error) {
 	file, err := os.Open(filePath)
 	if err != nil {
 		return "", err
